refactor: StatisticsTab을 model로 이동
